pyproject.toml
[tool.poetry]
name = "keyboard_project"
version = "0.1.0"
description = ""
authors = ["Mark Ionin <ioninmark91@gmail.com>"]
packages = [
  { include = "keyboard_project" },
]

[tool.poetry.dependencies]
python = "^3.9"
PyQt5 = "^5.15.6"
hidapi = "^0.12.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
x-keys = "keyboard_project.scripts.keyboard_control:main"